+++ src/connectedk8s/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 0.2.8
 ++++++
 * Added checks for proxy
+* Updated config dataplane endpoint to support other clouds
 
 0.2.7
 ++++++